Rename `io.spine.typehack` into `io.spine.tools.type`.
