+++ tools/plugin-base/src/main/java/io/spine/tools/type/MoreKnownTypes.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.typehack;
+package io.spine.tools.type;
 
 import io.spine.code.proto.FileSet;
 import io.spine.code.proto.TypeSet;
