+++ tools/plugin-base/src/main/java/io/spine/tools/type/package-info.java
@@ -29,7 +29,7 @@
 @Internal
 @CheckReturnValue
 @ParametersAreNonnullByDefault
-package io.spine.typehack;
+package io.spine.tools.type;
 
 import com.google.errorprone.annotations.CheckReturnValue;
 import io.spine.annotation.Internal;
